tests/wpt/tests/device-bound-session-credentials/refresh-does-not-send-challenge.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>DBSC refresh does not send back Sec-Session-Challenge</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="helper.js" type="module"></script>

<script type="module">
  import { expireCookie, waitForCookie, addCookieAndSessionCleanup, configureServer, setupShardedServerState, documentHasCookie } from "./helper.js";

  promise_test(async t => {
    await setupShardedServerState();
    const expectedCookieAndValue = "auth_cookie=abcdef0123";
    const expectedCookieAndAttributes = `${expectedCookieAndValue};Domain=${get_host_info().ORIGINAL_HOST};Path=/device-bound-session-credentials`;
    addCookieAndSessionCleanup(t);

    // Configure server to avoid sending back a challenge during refresh.
    configureServer({ refreshSendsChallenge: false });

    // Prompt starting a session, and wait until registration completes.
    const login_response = await fetch('login.py');
    assert_equals(login_response.status, 200);
    await waitForCookie(expectedCookieAndValue, /*expectCookie=*/true);

    // Confirm that a request has the cookie set.
    const auth_response = await fetch('verify_authenticated.py');
    assert_equals(auth_response.status, 200);

    // Trigger refresh and confirm that the cookie gets set again.
    expireCookie(expectedCookieAndAttributes);
    assert_false(documentHasCookie(expectedCookieAndValue));
    const auth_response_after_expiry = await fetch('verify_authenticated.py');
    assert_equals(auth_response_after_expiry.status, 200);
    assert_true(documentHasCookie(expectedCookieAndValue));
}, "Refresh does not send back Sec-Session-Challenge");
</script>
